requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
inflect==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.7.12
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
wincertstore==0.2
